requirements.txt
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.0.4
Pillow==8.1.0
PyJWT==2.0.1
whitenoise==5.2.0
